update hw selectors
