--- src/modules/hardwareWallet/store/devicesActions/hardwareWalletSelectors.test.js
@@ -1,23 +1,0 @@
-import {
-  selectActiveHardwareDevice,
-  selectActiveHardwareDeviceId,
-  selectHardwareDevices,
-} from '../selectors';
-
-describe('HardwareWallet selectors', () => {
-  const mockState = {
-    hardwareWallet: {
-      hardwareDevices: [{ deviceId: '1' }, { deviceId: '2' }],
-      activeHardwareDeviceId: '',
-    },
-  };
-  it('Should select ActiveHardwareDevice', async () => {
-    expect(selectActiveHardwareDevice(mockState)).toBeFalsy();
-  });
-  it('Should select ActiveHardwareDeviceId', async () => {
-    expect(selectActiveHardwareDeviceId(mockState)).toEqual('');
-  });
-  it('Should select HardwareDevices', async () => {
-    expect(selectHardwareDevices(mockState)).toEqual(mockState.hardwareWallet.hardwareDevices);
-  });
-});
