+++ src/modules/hardwareWallet/store/selectors/hwSelectors.js
